Part 25 - Confirmation dialog

+++ src/main/webapp/static/appjs/admin/category/list.controller.js
@@ -23,16 +23,23 @@ appCategory.controller('CategoryController', ['$scope', '$location', '$controlle
 
     $scope.deleteCategory = function(id){
         $controller('ModalController', {$scope: $scope})
-        CategoryService.delete({id: id}).$promise.then(
-            function onsuccess(){
-                $scope.open("success", "Category was deleted");
-                $scope.categories = CategoryService.query({page: page, size: size});
-            },
-            function onerror(response){
-                $scope.open("error", response.data.message);
-                $scope.categories = CategoryService.query({page: page, size: size});
+        $controller('ConfirmController', {$scope: $scope})
+
+        $scope.openConfirm('Are you sure you want to delete this entry', function(result){
+            if(result){
+                CategoryService.delete({id: id}).$promise.then(
+                    function onsuccess(){
+                        $scope.open("success", "Category was deleted");
+
+                    },
+                    function onerror(response){
+                        $scope.open("error", response.data.message);
+                    }
+                ).finally(function(){
+                        $scope.categories = CategoryService.query({page: page, size: size});
+                    })
             }
-        )
+        });
     }
 
     $scope.createCategory = function(){
